build: introduce scripts that create a bootable recovery.img from scratch
...relinquishing the need to repack an existing image.
The generated recovery.img is seemingly identical to the stock one, with the exception
of the recovery_dtbos being in a different order which makes no difference.

If the env var INSTALL_MAGISK is set to 1 when create_image.sh is invoked,
the recovery's ramdisk will be repacked with Magisk, saving you trips to Magisk Manager.
This requires magiskinit64 from a Magisk zip to be present.
(At this point in time, I would recommend taking it from 19.3 - avoid 20.0.)

When booted into Android proper or TWRP, flash_image.sh writes recovery.img to the
recovery partition on your phone and reboots it.

Signed-off-by: Faheem Pervez <trippin1@gmail.com>